Navigation준비작업
